% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_class.R
\name{Model}
\alias{Model}
\title{R6 Class Representing a Model Object}
\description{
A model object containing the model data and fitted models including
linear regression, decision tree and random forest models.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{dataframe}}{of features and target variable}

\item{\code{character, }}{selected algorithm}

\item{\code{character, }}{target variable column name}
}
\if{html}{\out{</div>}}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{dataframe}}{of features and target variable}

\item{\code{character, }}{selected algorithm}

\item{\code{character, }}{target variable column name}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{Model$new()}}
\item \href{#method-prepare_data}{\code{Model$prepare_data()}}
\item \href{#method-fit_model}{\code{Model$fit_model()}}
\item \href{#method-clone}{\code{Model$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Create a new Model object, load and prepare the model data
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Model$new(data_csv, target_variable)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{data_csv}}{character, the location of the model data}

\item{\code{target_variable}}{character, the name of the target column}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new `Model` object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-prepare_data"></a>}}
\if{latex}{\out{\hypertarget{method-prepare_data}{}}}
\subsection{Method \code{prepare_data()}}{
Import data, fill missing values and create features
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Model$prepare_data(data_file)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{data_file}}{character, file name of model data}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
dataframe of model features and target variable
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fit_model"></a>}}
\if{latex}{\out{\hypertarget{method-fit_model}{}}}
\subsection{Method \code{fit_model()}}{
Fit model with selected algorithm and features
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Model$fit_model(algo, features)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{algo}}{character, type of model to be used}

\item{\code{features}}{character, names of features to include}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
list of predicted values and R-Squared metric
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Model$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
